Split into one file per activity
