+++ connectors/src/connectors/google_drive/temporal/activities/common/constants.ts
@@ -1,0 +1,2 @@
+export const FILES_SYNC_CONCURRENCY = 10;
+export const FILES_GC_CONCURRENCY = 5;
