+++ connectors/src/connectors/google_drive/temporal/activities/common/types.ts
@@ -1,0 +1,5 @@
+export type LightGoogleDrive = {
+  id: string;
+  name: string;
+  isSharedDrive: boolean;
+};
