+++ connectors/src/connectors/google_drive/temporal/activities/garbage_collector_finished.ts
@@ -1,0 +1,11 @@
+import { ConnectorResource } from "@connectors/resources/connector_resource";
+import type { ModelId } from "@connectors/types";
+
+export async function garbageCollectorFinished(connectorId: ModelId) {
+  const connector = await ConnectorResource.fetchById(connectorId);
+  if (!connector) {
+    throw new Error(`Connector ${connectorId} not found`);
+  }
+
+  await connector.update({ lastGCTime: new Date() });
+}
